move LocalFeedItem to its own file
